backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
ortools>=9.10
openpyxl>=3.1
openai>=1.60
pydantic>=2.10
pydantic-settings>=2.7
pyyaml>=6.0
pytest>=8.0
httpx>=0.28
ruff>=0.9
